Changed the "tot_weights" variable in the updateGrade function to a double instead of an int. If the tot_weights was higher than 100, the scale variable at the bottom of the function would result in 0, because the result of integer division was being stored inside of it.

+++ LMS/LMS/Controllers/ProfessorController.cs
@@ -561,7 +561,7 @@ namespace LMS.Controllers
             // grade comes from enrolled, weight comes from assignment categories.
 
             double running_total = 0;
-            int tot_weights = 0;
+            double tot_weights = 0;
 
             var get_cats =
               from AC in db.AssignmentCategories
